Extract resource diagram
