+++ lib/csdl2openapi.js
@@ -10,6 +10,7 @@ const {
   nameParts,
   namespaceQualifiedName,
 } = require("./edm");
+const { resourceDiagram } = require("./diagram");
 
 //TODO
 // - Core.Example for complex types
@@ -69,16 +70,18 @@ module.exports.csdl2openapi = function (
 
   const model = new EDM();
   model.addDocument(csdl, messages);
-  //TODO: use directly from model, ideally remove
+
+  //TODO: use directly from model, ideally remove here
+  //TODO: only used in pathItemsForBoundOperations()
   const boundOverloads = model.boundOverloads;
+  //TODO: only used in schemasForStructuredType()
   const derivedTypes = model.derivedTypes;
+  //TODO: only used in ref()
   const namespace = model.namespace;
   const namespaceUrl = model.namespaceUrl;
-  const voc = model.voc;
 
-  const entityContainer = csdl.$EntityContainer
-    ? model.element(csdl.$EntityContainer)
-    : {};
+  const voc = model.voc;
+  const entityContainer = model.entityContainer;
   const keyAsSegment = entityContainer[voc.Capabilities.KeyAsSegmentSupported];
   const applySupported = entityContainer[voc.Aggregation.ApplySupported];
   const deepUpdate =
@@ -121,7 +124,7 @@ module.exports.csdl2openapi = function (
           serviceRoot +
           "/](" +
           serviceRoot.replace(/\(/g, "%28").replace(/\)/g, "%29") +
-          "/)") + (diagram ? resourceDiagram(csdl, entityContainer) : "");
+          "/)") + (diagram ? resourceDiagram(model) : "");
     return {
       title:
         entityContainer[voc.Core.Description] ||
@@ -136,201 +139,8 @@ module.exports.csdl2openapi = function (
   }
 
   /**
-   * Construct resource diagram using web service at https://yuml.me
-   * @param {object} csdl CSDL document
-   * @param {object} entityContainer Entity Container object
-   * @return {string} resource diagram
-   */
-  function resourceDiagram(csdl, entityContainer) {
-    let diagram = "";
-    let comma = "";
-    //TODO: make colors configurable
-    let color = {
-      resource: "{bg:lawngreen}",
-      entityType: "{bg:lightslategray}",
-      complexType: "",
-      external: "{bg:whitesmoke}",
-    };
-
-    Object.keys(csdl)
-      .filter((name) => isIdentifier(name))
-      .forEach((namespace) => {
-        const schema = csdl[namespace];
-        Object.keys(schema)
-          .filter(
-            (name) =>
-              isIdentifier(name) &&
-              ["EntityType", "ComplexType"].includes(schema[name].$Kind)
-          )
-          .forEach((typeName) => {
-            const type = schema[typeName];
-            diagram +=
-              comma +
-              (type.$BaseType
-                ? "[" +
-                  nameParts(type.$BaseType).name +
-                  (type.$Kind == "EntityType"
-                    ? color.entityType
-                    : color.complexType) +
-                  "]^"
-                : "") +
-              "[" +
-              typeName +
-              (type.$Kind == "EntityType"
-                ? color.entityType
-                : color.complexType) +
-              "]";
-            Object.keys(type)
-              .filter((name) => isIdentifier(name))
-              .forEach((propertyName) => {
-                const property = type[propertyName];
-                const targetNP = nameParts(property.$Type || "Edm.String");
-                if (
-                  property.$Kind == "NavigationProperty" ||
-                  targetNP.qualifier != "Edm"
-                ) {
-                  const target = model.element(property.$Type);
-                  const bidirectional =
-                    property.$Partner &&
-                    target &&
-                    target[property.$Partner] &&
-                    target[property.$Partner].$Partner == propertyName;
-                  // Note: if the partner has the same name then it will also be depicted
-                  if (!bidirectional || propertyName <= property.$Partner) {
-                    diagram +=
-                      ",[" +
-                      typeName +
-                      "]" +
-                      (property.$Kind != "NavigationProperty" ||
-                      property.$ContainsTarget
-                        ? "++"
-                        : bidirectional
-                        ? cardinality(target[property.$Partner])
-                        : "") +
-                      "-" +
-                      cardinality(property) +
-                      (property.$Kind != "NavigationProperty" || bidirectional
-                        ? ""
-                        : ">") +
-                      "[" +
-                      (target
-                        ? targetNP.name +
-                          (type.$Kind == "EntityType"
-                            ? color.entityType
-                            : color.complexType)
-                        : property.$Type + color.external) +
-                      "]";
-                  }
-                }
-              });
-            comma = ",";
-          });
-      });
-
-    Object.keys(entityContainer)
-      .filter((name) => isIdentifier(name))
-      .reverse()
-      .forEach((name) => {
-        const resource = entityContainer[name];
-        if (resource.$Type) {
-          diagram +=
-            comma +
-            "[" +
-            name +
-            "%20" +
-            color.resource +
-            "]" + // additional space in case entity set and type have same name
-            "++-" +
-            cardinality(resource) +
-            ">[" +
-            nameParts(resource.$Type).name +
-            "]";
-        } else {
-          if (resource.$Action) {
-            diagram += comma + "[" + name + color.resource + "]";
-            const overload = model
-              .element(resource.$Action)
-              .find((o) => !o.$IsBound);
-            diagram += overloadDiagram(name, overload);
-          } else if (resource.$Function) {
-            diagram += comma + "[" + name + color.resource + "]";
-            const overloads = model.element(resource.$Function);
-            if (overloads) {
-              const unbound = overloads.filter((o) => !o.$IsBound);
-              // TODO: loop over all overloads, add new source box after first arrow
-              diagram += overloadDiagram(name, unbound[0]);
-            }
-          }
-        }
-      });
-
-    if (diagram != "") {
-      diagram =
-        "\n\n## Entity Data Model\n![ER Diagram](https://yuml.me/diagram/class/" +
-        diagram +
-        ")\n\n### Legend\n![Legend](https://yuml.me/diagram/plain;dir:TB;scale:60/class/[External.Type" +
-        color.external +
-        "],[ComplexType" +
-        color.complexType +
-        "],[EntityType" +
-        color.entityType +
-        "],[EntitySet/Singleton/Operation" +
-        color.resource +
-        "])";
-    }
-
-    return diagram;
-
-    /**
-     * Diagram representation of property cardinality
-     * @param {object} typedElement Typed model element, e.g. property
-     * @param {boolean} one Explicitly represent to-1
-     * @return {string} cardinality
-     */
-    function cardinality(typedElement) {
-      return typedElement.$Collection
-        ? "*"
-        : typedElement.$Nullable
-        ? "0..1"
-        : "";
-    }
-
-    function overloadDiagram(name, overload) {
-      let diag = "";
-      if (overload.$ReturnType) {
-        const type = model.element(overload.$ReturnType.$Type || "Edm.String");
-        if (type) {
-          diag +=
-            "-" +
-            cardinality(overload.$ReturnType, true) +
-            ">[" +
-            nameParts(overload.$ReturnType.$Type).name +
-            "]";
-        }
-      }
-      for (const param of overload.$Parameter || []) {
-        const type = model.element(param.$Type || "Edm.String");
-        if (type) {
-          diag +=
-            comma +
-            "[" +
-            name +
-            color.resource +
-            "]in-" +
-            cardinality(param.$Type) +
-            ">[" +
-            nameParts(param.$Type).name +
-            "]";
-        }
-      }
-      return diag;
-    }
-  }
-
-  /**
    * Construct an array of Server Objects
-   * @param {object} csdl The CSDL document
-   * @param {object} options Optional parameters
+   * @param {string} serviceRoot Service root URL
    * @return {Array} The list of servers
    */
   function servers(serviceRoot) {
