+++ lib/diagram.js
@@ -1,0 +1,205 @@
+/**
+ * Resource diagram for an OData model using https://yuml.me
+ *
+ * Latest version: https://github.com/oasis-tcs/odata-openapi/blob/master/lib/edm.js
+ */
+
+const { isIdentifier, nameParts } = require("./edm");
+
+{
+  isIdentifier;
+}
+
+module.exports = { resourceDiagram };
+
+/**
+ * Construct resource diagram
+ * @param {object} model EDM for CSDL document
+ * @return {string} resource diagram
+ */
+function resourceDiagram(model) {
+  let diagram = "";
+  let comma = "";
+  //TODO: make colors configurable
+  let color = {
+    resource: "{bg:lawngreen}",
+    entityType: "{bg:lightslategray}",
+    complexType: "",
+    external: "{bg:whitesmoke}",
+  };
+
+  //TODO: for(const schema of model.schemas) // hide access details in model
+  Object.keys(model.csdl)
+    .filter((name) => isIdentifier(name))
+    .forEach((namespace) => {
+      const schema = model.csdl[namespace];
+      Object.keys(schema)
+        .filter(
+          (name) =>
+            isIdentifier(name) &&
+            ["EntityType", "ComplexType"].includes(schema[name].$Kind)
+        )
+        .forEach((typeName) => {
+          const type = schema[typeName];
+          diagram +=
+            comma +
+            (type.$BaseType
+              ? "[" +
+                nameParts(type.$BaseType).name +
+                (type.$Kind == "EntityType"
+                  ? color.entityType
+                  : color.complexType) +
+                "]^"
+              : "") +
+            "[" +
+            typeName +
+            (type.$Kind == "EntityType"
+              ? color.entityType
+              : color.complexType) +
+            "]";
+          Object.keys(type)
+            .filter((name) => isIdentifier(name))
+            .forEach((propertyName) => {
+              const property = type[propertyName];
+              const targetNP = nameParts(property.$Type || "Edm.String");
+              if (
+                property.$Kind == "NavigationProperty" ||
+                targetNP.qualifier != "Edm"
+              ) {
+                const target = model.element(property.$Type);
+                const bidirectional =
+                  property.$Partner &&
+                  target &&
+                  target[property.$Partner] &&
+                  target[property.$Partner].$Partner == propertyName;
+                // Note: if the partner has the same name then it will also be depicted
+                if (!bidirectional || propertyName <= property.$Partner) {
+                  diagram +=
+                    ",[" +
+                    typeName +
+                    "]" +
+                    (property.$Kind != "NavigationProperty" ||
+                    property.$ContainsTarget
+                      ? "++"
+                      : bidirectional
+                      ? cardinality(target[property.$Partner])
+                      : "") +
+                    "-" +
+                    cardinality(property) +
+                    (property.$Kind != "NavigationProperty" || bidirectional
+                      ? ""
+                      : ">") +
+                    "[" +
+                    (target
+                      ? targetNP.name +
+                        (type.$Kind == "EntityType"
+                          ? color.entityType
+                          : color.complexType)
+                      : property.$Type + color.external) +
+                    "]";
+                }
+              }
+            });
+          comma = ",";
+        });
+    });
+
+  Object.keys(model.entityContainer)
+    .filter((name) => isIdentifier(name))
+    .reverse()
+    .forEach((name) => {
+      const resource = model.entityContainer[name];
+      if (resource.$Type) {
+        diagram +=
+          comma +
+          "[" +
+          name +
+          "%20" +
+          color.resource +
+          "]" + // additional space in case entity set and type have same name
+          "++-" +
+          cardinality(resource) +
+          ">[" +
+          nameParts(resource.$Type).name +
+          "]";
+      } else {
+        if (resource.$Action) {
+          diagram += comma + "[" + name + color.resource + "]";
+          const overload = model
+            .element(resource.$Action)
+            .find((o) => !o.$IsBound);
+          diagram += overloadDiagram(name, overload);
+        } else if (resource.$Function) {
+          diagram += comma + "[" + name + color.resource + "]";
+          const overloads = model.element(resource.$Function);
+          if (overloads) {
+            const unbound = overloads.filter((o) => !o.$IsBound);
+            // TODO: loop over all overloads, add new source box after first arrow
+            diagram += overloadDiagram(name, unbound[0]);
+          }
+        }
+      }
+    });
+
+  if (diagram != "") {
+    diagram =
+      "\n\n## Entity Data Model\n![ER Diagram](https://yuml.me/diagram/class/" +
+      diagram +
+      ")\n\n### Legend\n![Legend](https://yuml.me/diagram/plain;dir:TB;scale:60/class/[External.Type" +
+      color.external +
+      "],[ComplexType" +
+      color.complexType +
+      "],[EntityType" +
+      color.entityType +
+      "],[EntitySet/Singleton/Operation" +
+      color.resource +
+      "])";
+  }
+
+  return diagram;
+
+  /**
+   * Diagram representation of property cardinality
+   * @param {object} typedElement Typed model element, e.g. property
+   * @param {boolean} one Explicitly represent to-1
+   * @return {string} cardinality
+   */
+  function cardinality(typedElement) {
+    return typedElement.$Collection
+      ? "*"
+      : typedElement.$Nullable
+      ? "0..1"
+      : "";
+  }
+
+  function overloadDiagram(name, overload) {
+    let diag = "";
+    if (overload.$ReturnType) {
+      const type = model.element(overload.$ReturnType.$Type || "Edm.String");
+      if (type) {
+        diag +=
+          "-" +
+          cardinality(overload.$ReturnType, true) +
+          ">[" +
+          nameParts(overload.$ReturnType.$Type).name +
+          "]";
+      }
+    }
+    for (const param of overload.$Parameter || []) {
+      const type = model.element(param.$Type || "Edm.String");
+      if (type) {
+        diag +=
+          comma +
+          "[" +
+          name +
+          color.resource +
+          "]in-" +
+          cardinality(param.$Type) +
+          ">[" +
+          nameParts(param.$Type).name +
+          "]";
+      }
+    }
+    return diag;
+  }
+}
