+++ lib/edm.js
@@ -4,12 +4,18 @@
  * Latest version: https://github.com/oasis-tcs/odata-openapi/blob/master/lib/edm.js
  */
 
+//TODO:
+// - move to odata-csdl repository?
+
 class EDM {
   //TODO: multi-document models
   addDocument(csdl, messages) {
     //TODO: should we really store the whole CSDL, or better extract its parts?
     this.csdl = csdl;
     this.preProcess(messages);
+    this.entityContainer = csdl.$EntityContainer
+      ? this.element(csdl.$EntityContainer)
+      : {};
   }
 
   //TODO: stronger encapsulation
@@ -139,8 +145,6 @@ class EDM {
         });
       });
   }
-
-  vocabularies() {}
 
   /**
    * Find model element by qualified name
